cors 임시 모두 허용

+++ src/main/java/project/ottshare/controller/WaitingUserController.java
@@ -2,9 +2,11 @@ package project.ottshare.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
+import project.ottshare.config.RedisConfig;
 import project.ottshare.dto.ottShareRoomDto.OttShareRoomRequestDto;
 import project.ottshare.dto.waitingUserDto.WaitingUserRequestDto;
 import project.ottshare.dto.waitingUserDto.WaitingUserResponseDto;
@@ -13,12 +15,14 @@ import project.ottshare.service.OttShareRoomService;
 import project.ottshare.service.WaitingUserService;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/waiting-user")
 @RequiredArgsConstructor
 @Slf4j
 public class WaitingUserController {
+
     private final WaitingUserService waitingUserService;
     private final OttShareRoomService ottShareRoomService;
 
@@ -40,16 +44,27 @@ public class WaitingUserController {
 
         waitingUserService.createWaitingUser(customUserDetails.userId(), waitingUserRequestDTO);
 
-        WaitingUserResponseDto leaderByOtt = waitingUserService.getLeaderByOtt(waitingUserRequestDTO.getOttType());
         List<WaitingUserResponseDto> members = waitingUserService.NonLeaderByOtt(waitingUserRequestDTO.getOttType());
+        if (members.isEmpty()){
+            return ResponseEntity.ok("멤버를 기다리는 중입니다.");
+        }
 
-        String ottId = leaderByOtt.getOttId();
-        String ottPassword = leaderByOtt.getOttPassword();
+        Optional<WaitingUserResponseDto> leaderByOtt = waitingUserService.getLeaderByOtt(waitingUserRequestDTO.getOttType());
+        if (leaderByOtt.isEmpty()){
+            return ResponseEntity.ok("방장이 없습니다.");
+        }
 
-        members.add(leaderByOtt);
+        WaitingUserResponseDto leader = leaderByOtt.get();
+        members.add(leader);
         waitingUserService.deleteAllWaitingUser(members);
 
-        OttShareRoomRequestDto ottShareRoomRequestDTO = ottShareRoomService.convertToOttShareRoom(members, ottId, ottPassword, waitingUserRequestDTO.getOttType());
+        OttShareRoomRequestDto ottShareRoomRequestDTO = ottShareRoomService.convertToOttShareRoom(
+                members,
+                leader.getOttId(),
+                leader.getOttPassword(),
+                waitingUserRequestDTO.getOttType()
+        );
+
         ottShareRoomService.createOttShareRoom(ottShareRoomRequestDTO);
 
         return ResponseEntity.ok("Room created successfully.");
